Tree : Top View With Python
You are given a pointer to the root of a binary tree. Print the top view of the binary tree. 
Top view means when you look the tree from the top the nodes you will see will be called the top view of the tree. See the example below. 
You only have to complete the function. 

Url:https://www.hackerrank.com/challenges/tree-top-view/problem